latte_gallery/pyproject.toml
[tool.poetry]
name = "latte-gallery"
version = "0.1.0"
description = ""
authors = ["arsuhinars <arsuhinars@gmail.com>"]
readme = "README.md"

[tool.mypy]
plugins = "pydantic.mypy"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
psycopg = {extras = ["binary"], version = "^3.2.3"}
pydantic-settings = "^2.6.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
